// src/app/core/constans/enums.ts
export enum IconIds {
  TRASH = 'trash',
  PENCIL_SQUARE = 'pencilSquare',
  X_CIRCLE = 'xCircle',
  CHECK_CIRCLE = 'checkCircle',
  EYE = 'eye',
  EYE_SLASH = 'eyeSlash',
  BAN = 'ban',
  EXCLAMATION_CIRCLE = 'exclamationCircle',
  INFO_CIRCLE = 'infoCircle',
}

export enum TableOperationEnum {
  EDIT = 'edit',
  DELETE = 'delete',
  CHECK = 'checkMark',
  CANCEL = 'cancel',
  SEE = 'see',
}